refactor(web): reorganize log page
- Split up table and route
- Renamed "logs" to "log"
- Added "Not Found" indication
